requirements.txt
aioredis==1.2.0
asgiref==3.2.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.2
Automat==0.7.0
bcrypt==3.1.7
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
constantly==15.1.0
cryptography==2.7
daphne==2.3.0
Django==2.2.3
django-cors-headers==3.0.2
django-redis==4.10.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
dwebsocket==0.5.10
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
msgpack==0.6.1
paramiko==2.6.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
python-masscan==0.1.3
pytz==2019.1
redis==3.3.2
six==1.12.0
sqlparse==0.3.0
Twisted==19.2.1
txaio==18.8.1
xlwt==1.3.0
zope.interface==4.6.0
